clean up file naming
